+++ src/main/java/com/keldorn/todocorejavaspringsolution/dto/todo/TodoRequest.java
@@ -4,5 +4,5 @@ import com.keldorn.todocorejavaspringsolution.domain.enums.Priority;
 
 import java.time.LocalDateTime;
 
-public record TodoRequest(String title, String description, LocalDateTime dueDate, Boolean completed, Priority priority, int userId) {
+public record TodoRequest(String title, String description, LocalDateTime dueDate, Boolean completed, Priority priority, Long userId) {
 }
