+++ src/main/java/com/keldorn/todocorejavaspringsolution/dto/todo/TodoResponse.java
@@ -2,6 +2,6 @@ package com.keldorn.todocorejavaspringsolution.dto.todo;
 
 import java.time.LocalDateTime;
 
-public record TodoResponse(int todoId, String title, String description, LocalDateTime dueDate,
-                           boolean completed, int priority, int userId) {
+public record TodoResponse(Long todoId, String title, String description, LocalDateTime dueDate,
+                           boolean completed, int priority, Long userId) {
 }
